pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=2.0.0,<3.0.0"]

[dependency-groups]
dev = [
  "pylint-pydantic (>=0.4.1,<0.5.0)",
  "pytest-asyncio (>=1.3.0,<2.0.0)"
]

[project]
authors = [
  {email = "zhulikovnikita884@gmail.com", name = "nikitazhulikov"}
]
dependencies = [
  "amocrm-api (>=2.6.1,<3.0.0)",
  "black (>=25.11.0,<26.0.0)",
  "dotenv (>=0.9.9,<0.10.0)",
  "flake8 (>=7.3.0,<8.0.0)",
  "flake8-pyproject (>=1.2.3,<2.0.0)",
  "google-auth (>=2.43.0,<3.0.0)",
  "gspread (>=6.2.1,<7.0.0)",
  "httpx (>=0.28.1,<0.29.0)",
  "isort (>=7.0.0,<8.0.0)",
  "mypy (>=1.18.2,<2.0.0)",
  "pydantic-settings (>=2.12.0,<3.0.0)",
  "pylint (>=4.0.2,<5.0.0)",
  "pytest (>=9.0.0,<10.0.0)",
  "tenacity (>=9.1.2,<10.0.0)",
  "toml-sort (>=0.24.3,<0.25.0)"
]
description = ""
name = "amocrm-events-monitor"
readme = "README.md"
requires-python = ">=3.13"
version = "0.1.0"

[tool.black]
line-length = 128
preview = true
target-version = ["py312"]

[tool.flake8]
exclude = [
  "./.mypy_cache/*,",
  "./.venv/*",
  ".DS_Store",
  ".git",
  ".tox",
  "__pycache__"
]
ignore = ["E203", "E501", "E704", "W503"]
max-line-length = 128
plugins = ["flake8-pyproject"]

[tool.isort]
profile = "black"

[tool.mypy]
check_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_calls = true
disallow_untyped_defs = true
follow_imports_for_stubs = true
ignore_missing_imports = true
local_partial_types = true
namespace_packages = true
pretty = true
show_error_context = true
warn_redundant_casts = true
warn_return_any = true
warn_unused_configs = true
warn_unused_ignores = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
  "integration: marks tests as integration tests requiring real external services (deselect with '-m \"not integration\"')"
]
